agents/pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "elasticsearch>=9.1.1",
    "fastapi[standard]>=0.120.0",
    "langchain>=1.0.2",
    "langchain-community>=0.4",
    "langchain-openai>=1.0.1",
    "langchain-perplexity>=1.0.0",
    "python-dotenv>=1.1.1",
    "unstructured[pdf]>=0.18.15",
]
